grapegram_server/pyproject.toml
[tool.poetry]
name = "grapegram-server"
version = "0.1.0"
description = ""
authors = ["Blacgolyb <andrejomelnickij@gmail.com>"]
readme = "README.md"
packages = [{include = "grapegram_server"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
asyncpg = "^0.27.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.15"}
python-jose = "^3.3.0"
envparse = "^0.2.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
pymongo = "^4.3.3"
cryptography = "^40.0.2"
ngrok = "^0.8.0"
websockets = "^11.0.3"
bcrypt = "^4.0.1"
alembic = "^1.11.1"
psycopg2 = "^2.9.6"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
